% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changer.R
\name{changer.RasterBrick}
\alias{changer.RasterBrick}
\title{Changepoint detection for stars objects}
\usage{
\method{changer}{RasterBrick}(x, dates, method = "prophet", FUTURE = FALSE, ...)
}
\arguments{
\item{x}{raster. Spatiotemporal cube in the form of a raster stack or raster brick.}

\item{dates}{date (or similar). Vector of dates, 1 per layer in x.}

\item{method}{character. One of prophet, ...}

\item{FUTURE}{logical. Controls whether the calculation should be done in parallel via the future package.}

\item{...}{arguments passed to the underlying method.}
}
\description{
Changepoint detection for stars objects
}
\details{
Warning. raster converts to an array internally and therefore involves a copy.
}
