% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-changePoints.R
\name{compareWithWAIC}
\alias{compareWithWAIC}
\title{Compare models using waic}
\usage{
compareWithWAIC(fit)
}
\arguments{
\item{fit}{A list of mcp model fits}
}
\value{
A loo model comparison object
}
\description{
Compare models using waic
}
