% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-dataModule.R
\name{dataUI}
\alias{dataUI}
\alias{dataServer}
\title{UI function of data module}
\usage{
dataUI(id)

dataServer(id, transformations = list(), path = NULL)
}
\arguments{
\item{id}{module id}

\item{transformations}{list of transformations to apply to the dataset}

\item{path}{path to the example data file, e.g. \code{file.path("data", "example.csv")}}
}
\description{
UI function of data module

Server function of data module
}
