\name{getID}
\alias{getID}
\title{Get the ID for a petition}
\usage{
  getID(url)
}
\arguments{
  \item{url}{The URL of the petition on www.change.org}
}
\value{
  a list with "petition_id" component that has the petition
  id, and "result" component with the status string.
}
\description{
  Get the ID for a petition
}
\author{
  Barry Rowlingson
}

