% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-mcpModelingModule.R
\name{mcpDataUI}
\alias{mcpDataUI}
\alias{mcpDataServer}
\title{UI function of mcpData module}
\usage{
mcpDataUI(id)

mcpDataServer(id, file_data)
}
\arguments{
\item{id}{module id}

\item{file_data}{reactive file data}
}
\description{
UI function of mcpData module

Server function of mcpData module
}
