% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-mcpFormulasModule.R
\name{mcpFormulasUI}
\alias{mcpFormulasUI}
\alias{mcpFormulasServer}
\title{MCP Formulas UI}
\usage{
mcpFormulasUI(id)

mcpFormulasServer(id, uploaded_matrices = reactiveValues())
}
\arguments{
\item{id}{The module id}

\item{uploaded_matrices}{(reactive) values of uploaded segments and priors matrices}
}
\description{
MCP Formulas UI

MCP Formulas Server
}
