% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-mcpModelingModule.R
\name{mcpModelSummaryUI}
\alias{mcpModelSummaryUI}
\alias{mcpModelSummaryServer}
\title{MCP Model Summary UI}
\usage{
mcpModelSummaryUI(id)

mcpModelSummaryServer(id, mcpData, formulasAndPriors, mcpFitList)
}
\arguments{
\item{id}{The module id}

\item{mcpData}{The reactive mcp data}

\item{formulasAndPriors}{The reactive formulas and priors}

\item{mcpFitList}{The reactive mcp fit list}
}
\description{
MCP Model Summary UI

MCP Model Summary Server
}
