% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-mcpModelingModule.R
\name{mcpOutUI}
\alias{mcpOutUI}
\alias{mcpOutServer}
\title{MCP Output UI}
\usage{
mcpOutUI(id, title, outFUN = verbatimTextOutput, showWidth = FALSE)

mcpOutServer(
  id,
  formulasAndPriors,
  mcpData,
  mcpFitList,
  mcpModelName,
  mcpOutFUN,
  renderFUN = renderPrint
)
}
\arguments{
\item{id}{The module id}

\item{title}{The title of the output}

\item{outFUN}{The output function, either verbatimTextOutput or plotOutput}

\item{showWidth}{Show the width slider}

\item{formulasAndPriors}{The reactive formulas and priors}

\item{mcpData}{The reactive mcp data}

\item{mcpFitList}{The reactive mcp fit list}

\item{mcpModelName}{The reactive mcp model name}

\item{mcpOutFUN}{The output function, either "summary", "waic" or "plot"}

\item{renderFUN}{The render function, either renderPrint or renderPlot}
}
\description{
MCP Output UI

MCP Output Server
}
