% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixel_prophet.R
\name{pixel_prophet}
\alias{pixel_prophet}
\title{A function to run the prophet algorithm by pixel}
\usage{
pixel_prophet(x, dates, ...)
}
\arguments{
\item{x}{numeric vector}

\item{dates}{vector of dates, with equal length to x}

\item{...}{arguments passed to `prophet`}
}
\description{
A function to run the prophet algorithm by pixel
}
