% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-changePoints.R
\name{runMcp}
\alias{runMcp}
\title{Run mcp model}
\usage{
runMcp(lists, ...)
}
\arguments{
\item{lists}{A list of lists containing the segmented formulas and priors, output of \code{setFormulasAndPriors()}}

\item{...}{Additional arguments to pass to mcp}
}
\value{
A list of mcp model fits
}
\description{
Run mcp model
}
