% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-changePointsModule.R
\name{changePointsUI}
\alias{changePointsUI}
\alias{changePointsServer}
\title{Change Points UI}
\usage{
changePointsUI(id)

changePointsServer(id, file_data)
}
\arguments{
\item{id}{The module id}

\item{file_data}{The reactive file data}
}
\description{
UI of the module

Server function of the module
}
