% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changer.R
\name{changer}
\alias{changer}
\title{Identify change points in spatiotemporal cubes}
\usage{
changer(x, ...)
}
\arguments{
\item{x}{A spatiotemporal cube with a class of `stars` (and in the future, `raster`). Assumes dimensions are (x,y,time)}

\item{...}{arguments tbd.}
}
\description{
Identify change points in spatiotemporal cubes
}
