% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-dataModule.R
\name{dataUI}
\alias{dataUI}
\alias{dataServer}
\title{UI function of data module}
\usage{
dataUI(id)

dataServer(id)
}
\arguments{
\item{id}{module id}
}
\description{
UI function of data module

Server function of data module
}
