% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-mcpModelingModule.R
\name{mcpDataUI}
\alias{mcpDataUI}
\alias{mcpDataServer}
\title{UI function of mcpData module}
\usage{
mcpDataUI(id)

mcpDataServer(id, file_data, mcp_columns = c(x = "", y = ""))
}
\arguments{
\item{id}{module id}

\item{file_data}{reactive file data}

\item{mcp_columns}{(character) a character vector containing the default x and y columns for MCP modelling.
The first entry corresponds to the x column name, and the second entry corresponds to the y column name.}
}
\description{
UI function of mcpData module

Server function of mcpData module
}
